refactor: input vlaues
